[3879] changes as per PR comments
