+++ stream-chat-android-core/src/main/java/io/getstream/chat/android/core/internal/InternalStreamChatApi.kt
@@ -16,6 +16,13 @@
 
 package io.getstream.chat.android.core.internal
 
+@Target(
+    AnnotationTarget.CLASS,
+    AnnotationTarget.PROPERTY,
+    AnnotationTarget.CONSTRUCTOR,
+    AnnotationTarget.FUNCTION,
+    AnnotationTarget.TYPEALIAS
+)
 @RequiresOptIn(
     message = "This is internal API for the Stream Chat libraries. Do not depend on this API in your own client code.",
     level = RequiresOptIn.Level.ERROR
