+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/offline/sync/internal/SyncManager.kt
@@ -68,8 +68,8 @@ import java.util.concurrent.atomic.AtomicReference
 private const val QUERIES_TO_RETRY = 3
 
 /**
- * This class is responsible to sync messages, reactions and channel data. It tries to sync then, if necessary, when connection
- * is reestabilished or when a health check even happens.
+ * This class is responsible to sync messages, reactions and channel data. It tries to sync then, if necessary,
+ * when connection is reestablished or when a health check event happens.
  */
 @Suppress("LongParameterList", "TooManyFunctions", "TooGenericExceptionCaught")
 internal class SyncManager(
